move IpcGraph to ipc module
