+++ modules/ipc/tests/CMakeLists.txt
@@ -36,3 +36,10 @@ define_module_test(
   PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
   PUBLIC_LINK_LIBS hephaestus::random hephaestus::types_proto
 )
+
+define_module_test(
+  NAME ipc_graph_tests
+  SOURCES ipc_graph_tests.cpp
+  PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
+  PUBLIC_LINK_LIBS hephaestus::types hephaestus::types_proto
+)
