+++ modules/websocket_bridge/include/hephaestus/websocket_bridge/bridge.h
@@ -10,13 +10,13 @@
 #include <span>
 #include <string>
 
+#include <hephaestus/ipc/zenoh/ipc_graph.h>
 #include <hephaestus/ipc/zenoh/liveliness.h>
 #include <hephaestus/ipc/zenoh/raw_subscriber.h>
 #include <hephaestus/ipc/zenoh/session.h>
 #include <hephaestus/serdes/type_info.h>
 
 #include "hephaestus/ipc/ipc_entity_manager.h"
-#include "hephaestus/ipc/ipc_graph.h"
 #include "hephaestus/websocket_bridge/bridge_config.h"
 #include "hephaestus/websocket_bridge/bridge_state.h"
 #include "hephaestus/websocket_bridge/utils/ws_protocol.h"
@@ -61,6 +61,13 @@ private:
   ///////////////////
   // IPC Interface //
   ///////////////////
+
+  using IpcGraph = ipc::zenoh::IpcGraph;
+  using IpcGraphState = ipc::zenoh::IpcGraphState;
+  using IpcGraphConfig = ipc::zenoh::IpcGraphConfig;
+  using TopicsToTypeMap = ipc::zenoh::TopicsToTypeMap;
+  using TopicsToServiceTypesMap = ipc::zenoh::TopicsToServiceTypesMap;
+  using TopicToSessionIdMap = ipc::zenoh::TopicToSessionIdMap;
 
   std::unique_ptr<IpcGraph> ipc_graph_;
